requirements.txt
PyPDF2
faiss-cpu
streamlit
transformers
google-generativeai
sentence_transformers
